agents/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-google-genai==2.0.8
pydantic==2.10.4
httpx==0.28.1
